create chain id package
