+++ x/launch/types/message_edit_chain.go
@@ -3,6 +3,7 @@ package types
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/tendermint/spn/pkg/chainid"
 )
 
 var _ sdk.Msg = &MsgEditChain{}
@@ -53,7 +54,7 @@ func (msg *MsgEditChain) ValidateBasic() error {
 	}
 
 	if msg.GenesisChainID != "" {
-		if _, _, err := ParseGenesisChainID(msg.GenesisChainID); err != nil {
+		if _, _, err := chainid.ParseGenesisChainID(msg.GenesisChainID); err != nil {
 			return sdkerrors.Wrapf(ErrInvalidGenesisChainID, msg.GenesisChainID)
 		}
 	}
